WassersteinDeepEmbeddings/pyproject.toml
[tool.poetry]
name = "jax_dwe"
version = "0.2.1"
description = "CNN based embeddings for Wasserstein Distances"
authors = ["Anon"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flax = "0.7.5"
ott-jax = "0.4.4"
clu = "0.0.10"
tqdm = "^4.66.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
